+++ src/pas/plugins/kimug/browser/view.py
@@ -1,13 +1,23 @@
 from pas.plugins.kimug.utils import get_keycloak_users
 from pas.plugins.kimug.utils import migrate_plone_user_id_to_keycloak_user_id
 from pas.plugins.kimug.utils import set_oidc_settings
+from pas.plugins.kimug.utils import get_keycloak_users_from_oidc
+from pas.plugins.kimug.utils import add_keycloak_users_to_plone
 from plone import api
 from Products.Five.browser import BrowserView
+from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 
 import logging
 
 
-logger = logging.getLogger("pas.plugins.kimug.view")
+# from zope.interface import Interface
+# from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+
+
+logger = logging.getLogger("pas.plugins.kimug")
+
+# class IMyView(Interface):
+#     """Marker Interface for IMyView"""
 
 
 class MigrationView(BrowserView):
@@ -27,3 +37,16 @@ class SetOidcSettingsView(BrowserView):
         api.portal.show_message("OIDC settings configured successfully", self.request)
         logger.info("OIDC settings configured successfully")
         self.request.response.redirect(self.context.absolute_url())
+
+
+class KeycloakUsersView(BrowserView):
+    # If you want to define a template here, please remove the template attribute from
+    # the configure.zcml registration of this view.
+    # template = ViewPageTemplateFile('my_view.pt')
+    index = ViewPageTemplateFile("users.pt")
+
+    def __call__(self):
+        keycloak_users = get_keycloak_users_from_oidc()
+        added_users = add_keycloak_users_to_plone(keycloak_users)
+        api.portal.show_message(f"{added_users} Keycloak users imported", self.request)
+        return self.index()